src/main/java/com/ade/open/api/server/service/SpringBeanFactory.java
package com.ade.open.api.server.service;

import com.ade.open.api.server.dao.TOpenSecretDao;
import org.springframework.beans.BeansException;
import org.springframework.context.ApplicationContext;
import org.springframework.context.ApplicationContextAware;
import org.springframework.stereotype.Component;

/**
 * Created by liyang on 17-12-15.
 */
@Component
public class SpringBeanFactory implements ApplicationContextAware {

    private static ApplicationContext applicationContext; // Spring应用上下文环境

    public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
        SpringBeanFactory.applicationContext = applicationContext;
    }

    public static ApplicationContext getApplicationContext() {
        return applicationContext;
    }

    public static TOpenSecretDao getTOpenSecretDao() {
        return applicationContext.getBean(TOpenSecretDao.class);
    }

}
